requirements.txt
azure-communication-email==1.0.0
azure-kusto-data==3.1.3
